# perf: replace `identical(x, TRUE)` with faster `isTRUE(x)`/`isFALSE(x)`