@s-fleck 2025-04-11
Hi! great input, thanks. I actually thought it would be the other way round performance wise... i'll need a bit to review this but i'll merge it in a week or so!